Fix typo (copy-paste error?) in last question
In the case of an even number of points, after a 180° turn, the pivot point must be the one to the right of the initial pivot, else the number of blue vs brown points would have changed: the initial pivot would still be brown, but every other point changed color.
Since we had one less brown non-pivot point, we'd have one less blue non-pivot point now, so two more blue than brown points. 
If we're considering 360° instead, then we should be fine, since whatever was left of the line initially is left again, same for right.