Update index.mdx (#256)
I can't see (okay, now I see it --I wonder why ctrl + f  doesn't find it?) the bit I want to change here, that I read in the article (https://www.3blue1brown.com/lessons/neural-networks#title). I had to click on a 6 in a blue dot to see it in the article, but I have no idea how to find it here at GitHub. That is a problem in itself that should be fixed, isn't it? This is the pop up:

"These days, due to the demand for more machine learning and more powerful networks, there have been huge leaps and bounds in specialized hardware which is aimed at making matrix multiplication much faster, for example, Google's "Tensor processing unit", or TPU. 

Half the time when you hear a company describe something like a "new neural architecture" meant for more powerful AI, what they really mean when you look under the hood is that they're multiplying matrices more quickly.

In fairness, this hardware often does more than just multiply matrices, but that's a huge part of it. And if you understand the above section, you understand why."

Although not exactly badly written, it is mediocre compared the rest of the article which is really well written, I mean *stellar*. Beautiful. This one paragraph therefore sticks out like a sore thumb for me, because it contains two small flaws. 

So I thought I'd "noticed a mistake" and so I clicked on "Submit a correction at GitHub". It's very satisfying to make something almost perfect even closer to being perfect.

The first flaw is the use of the phrase "huge leaps and bounds". "by leaps and bounds" is an idiom and it can just about be truncated to "leaps and bounds" without looking bad, but "huge leaps and bounds" is jarring and suggests that the writer is not a native speaker. The other flaw is the use of "that's a huge part of it". Is it most of it? If so, you should say so. If not, you should say how big a part of it it is. Or just have:: 

"Half the time when you hear a company describe something like a "new neural architecture" meant for more powerful AI, what they really mean when you look under the hood is that they're multiplying matrices more quickly, although, in fairness, this type of hardware often does more than just multiply matrices."

and,

"Lately, due to the demand for more machine learning and more powerful networks, there is specialized hardware (that is improving by leaps and bounds) that can multiply matrices much faster, for example, Google's "Tensor Processing Unit", or TPU.